+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## implemented but not yet released
 - use a new rendering hook, eliminating graphical glitches near water
-- add LSDPreRenderEvent.partialTick property
+- add getPartialTick method to LSDPreRenderEvent and MinecraftAccess.
 - improved Trident animation library integration: built-in interpolators for
   types in `libshapedraw.primitive`
 - add several convenience methods